5 Define metrics: fft and laplacian and sketch image processing configuration model

+++ src/main/java/com/rgosiewski/frameiq/repository/management/ProjectRepository.java
@@ -6,8 +6,10 @@
 package com.rgosiewski.frameiq.repository.management;
 
 import com.rgosiewski.frameiq.repository.enums.FileExtensions;
+import com.rgosiewski.frameiq.repository.exception.InvalidProjectPathException;
 import com.rgosiewski.frameiq.repository.exception.NewRepositoryDirectoryException;
 import com.rgosiewski.frameiq.repository.exception.NewRepositoryFileException;
+import com.rgosiewski.frameiq.repository.exception.NonExistingProjectException;
 import org.apache.logging.log4j.Level;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
@@ -15,6 +17,7 @@ import org.springframework.stereotype.Service;
 
 import java.io.File;
 import java.io.FileOutputStream;
+import java.nio.file.Path;
 
 @Service
 class ProjectRepository {
@@ -27,20 +30,21 @@ class ProjectRepository {
 
     public void createProjectDir(String projectName) {
         logger.log(Level.INFO, "Creating new project: {}", projectName);
-        File projectDir;
-        try {
-            projectDir = getProjectDir(projectName);
-        } catch (Exception e) {
-            throw new NewRepositoryDirectoryException(projectName);
-        }
-        if (!projectDir.mkdir()) {
+        File projectDir = getProjectDir(projectName);
+        if (!projectDir.exists() && !projectDir.mkdir()) {
             throw new NewRepositoryDirectoryException(projectName);
         }
         logger.log(Level.INFO, "Project created: {}", projectName);
     }
 
     public File getProjectDir(String projectName) {
-        return getRepositoryRoot().toPath().resolve(projectName).toFile();
+
+        try {
+            Path projectPath = getRepositoryRoot().toPath().resolve(projectName);
+            return projectPath.toFile();
+        } catch (Exception e) {
+            throw new InvalidProjectPathException(projectName);
+        }
     }
 
     private File getRepositoryRoot() {
@@ -50,6 +54,9 @@ class ProjectRepository {
     public void addFileToProject(byte[] fileContent, String fileName, FileExtensions extension, String projectName) {
         logger.log(Level.INFO, "Creating new {} file: {} in the project: {}", extension, fileName, projectName);
         File projectDir = getProjectDir(projectName);
+        if ( !projectDir.exists()) {
+            throw new NonExistingProjectException(projectName);
+        }
         File newFile = projectDir.toPath().resolve(fileName + "." + extension.getExtension()).toFile();
         try (FileOutputStream fileOutputStream = new FileOutputStream(newFile)) {
             fileOutputStream.write(fileContent);
